pyproject.toml
[project]
name = "study-fastapi"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "qe-present", email = "2664481691@qq.com" }
]
dependencies = [
    "fastapi[standard]>=0.125.0",
    "uvicorn>=0.35.0",
    "asyncpg>=0.31.0",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.27",
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/study_fastapi"]
